pyproject.toml
[tool.poetry]
name = "soccer-player-tracking"
version = "0.1.0"
description = "Soccer player detection and counting using YOLOv8"
authors = ["George Drakoulas <geo.drakoulas@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
ultralytics = ">=8.0.0"
opencv-python = ">=4.8.0"
PyYAML = ">=6.0"
numpy = ">=1.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
